beginning of stop segment work, tag inlining and loops

+++ _shared_utils/shared_utils/utils.py
@@ -7,6 +7,7 @@ import shutil
 from pathlib import Path
 from typing import Union
 
+import dask_geopandas as dg
 import fsspec
 import geopandas as gpd
 import requests
@@ -22,7 +23,7 @@ def sanitize_file_path(file_name: str) -> str:
     return str(Path(file_name).stem)
 
 
-def geoparquet_gcs_export(gdf: gpd.GeoDataFrame, gcs_file_path: str, file_name: str):
+def geoparquet_gcs_export(gdf: Union[gpd.GeoDataFrame, dg.GeoDataFrame], gcs_file_path: str, file_name: str):
     """
     Save geodataframe as parquet locally,
     then move to GCS bucket and delete local file.
@@ -34,12 +35,17 @@ def geoparquet_gcs_export(gdf: gpd.GeoDataFrame, gcs_file_path: str, file_name: 
                 Filename, with or without .parquet.
     """
     file_name_sanitized = sanitize_file_path(file_name)
-    gdf.to_parquet(f"./{file_name_sanitized}.parquet")
-    fs.put(
-        f"./{file_name_sanitized}.parquet",
-        f"{gcs_file_path}{file_name_sanitized}.parquet",
-    )
-    os.remove(f"./{file_name_sanitized}.parquet")
+
+    if isinstance(gdf, dg.GeoDataFrame):
+        gdf.to_parquet(f"{gcs_file_path}{file_name_sanitized}", overwrite=True)
+
+    else:
+        gdf.to_parquet(f"./{file_name_sanitized}.parquet")
+        fs.put(
+            f"./{file_name_sanitized}.parquet",
+            f"{gcs_file_path}{file_name_sanitized}.parquet",
+        )
+        os.remove(f"./{file_name_sanitized}.parquet")
 
 
 def geojson_gcs_export(
